refactor[FE]: Refactor types
