+++ frontend/prettier.config.js
@@ -1,10 +1,10 @@
 module.exports = {
   printWidth: 80,
-  trailingComma: "es5",
+  trailingComma: 'es5',
   tabWidth: 2,
   semi: true,
   singleQuote: true,
   bracketSameLine: false,
-  tailwindConfig: "./tailwind.config.js",
+  tailwindConfig: './tailwind.config.js',
   plugins: [],
-}
+};
