+++ frontend/src/components/molecule/ChannelList.tsx
@@ -1,5 +1,5 @@
 import ChannelItem from 'components/molecule/ChannelItem';
-import { ChannelType } from 'types/channelType';
+import { ChannelType } from 'types';
 
 interface Props {
   styles: { readonly [key: string]: string };
