+++ frontend/src/components/molecule/ChannelUserItem.tsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
 import ProfileImage from 'components/atoms/ProfileImage';
 import Button from 'components/atoms/Button';
-import { ChannelUserType, UserRole } from 'types/channelUserType';
+import { ChannelUserType, CHANNEL_USER_STATUS, USER_ROLES } from 'types';
 
 interface Props {
   changeRole: (arg: any) => void;
@@ -21,12 +21,13 @@ export default function ChannelUserItem({
   myUser,
 }: Props) {
   const isSelf = myUser?.id === user.id;
-  const amIOwner = myUser?.role === UserRole.owner;
+  const amIOwner = myUser?.role === USER_ROLES.OWNER;
 
   const handleClickRole = () => {
     changeRole({
       userId: user.id,
-      role: user.role === UserRole.admin ? 'normal' : 'admin',
+      role:
+        user.role === USER_ROLES.ADMIN ? USER_ROLES.NORMAL : USER_ROLES.ADMIN,
     });
   };
 
@@ -49,8 +50,8 @@ export default function ChannelUserItem({
       </Link>
       <Link to={`/profile/${user.id}`}>
         <span>
-          {user.role === UserRole.owner && `🕶`}
-          {user.role === UserRole.admin && `👓`}
+          {user.role === USER_ROLES.OWNER && `🕶`}
+          {user.role === USER_ROLES.ADMIN && `👓`}
           {user.nickname}
         </span>
       </Link>
@@ -60,7 +61,7 @@ export default function ChannelUserItem({
 
           {amIOwner && (
             <>
-              {user.role === UserRole.admin ? (
+              {user.role === USER_ROLES.ADMIN ? (
                 <Button type="button" onClick={handleClickRole}>
                   관리자 해제
                 </Button>
@@ -69,23 +70,27 @@ export default function ChannelUserItem({
                   관리자 임명
                 </Button>
               )}
-              {user.role !== UserRole.owner && (
+              {user.role !== USER_ROLES.OWNER && (
                 <>
                   <Button
                     type="button"
-                    value="mute"
+                    value={CHANNEL_USER_STATUS.MUTE}
                     onClick={handleClickStatus}
                   >
                     조용히
                   </Button>
                   <Button
                     type="button"
-                    value="kick"
+                    value={CHANNEL_USER_STATUS.KICK}
                     onClick={handleClickStatus}
                   >
                     내보내기
                   </Button>
-                  <Button type="button" value="ban" onClick={handleClickStatus}>
+                  <Button
+                    type="button"
+                    value={CHANNEL_USER_STATUS.BAN}
+                    onClick={handleClickStatus}
+                  >
                     차단하기
                   </Button>
                 </>
