+++ frontend/src/components/molecule/SearchCombobox.tsx
@@ -1,9 +1,7 @@
 import { Fragment, useEffect, useState } from 'react';
 import { Combobox } from '@headlessui/react';
 import ProfileImage from 'components/atoms/ProfileImage';
-import { UserType } from 'types/userType';
-import { ChannelUserType } from 'types/channelUserType';
-
+import { UserType, ChannelUserType } from 'types';
 interface Props {
   placeholder?: string;
   dataList?: UserType[];
