+++ frontend/src/components/organisms/Achievements/AchievementList.tsx
@@ -1,5 +1,5 @@
 import AchievementItem from 'components/organisms/Achievements/AchievementItem';
-import { AchievementType } from 'types/achievementType';
+import { AchievementType } from 'types';
 import styles from 'assets/styles/Achievement.module.css';
 
 interface Props {
