+++ frontend/src/components/organisms/ChannelBrowse.tsx
@@ -1,5 +1,5 @@
 import ChannelList from 'components/molecule/ChannelList';
-import { ChannelType } from 'types/channelType';
+import { ChannelType } from 'types';
 import styles from 'assets/styles/ChannelLobby.module.css';
 
 interface Props {
