+++ frontend/src/components/organisms/SearchResult/SearchResultItem.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 import ProfileImage from 'components/atoms/ProfileImage';
-import { UserType } from 'types/userType';
+import { UserType } from 'types';
 import styles from 'assets/styles/SearchResult.module.css';
 
 interface Props {
