+++ frontend/src/components/organisms/SearchResult/index.tsx
@@ -1,5 +1,5 @@
 import SearchResultList from 'components/organisms/SearchResult/SearchResultList';
-import { UserType } from 'types/userType';
+import { UserType } from 'types';
 
 interface Props {
   users: UserType[];
