+++ frontend/src/components/organisms/SettingProfile.tsx
@@ -3,8 +3,7 @@ import { API_PREFIX } from 'api/api.v1';
 import TextInputWithMessage from 'components/molecule/TextInputWithMessage';
 import FileInputWithImage from 'components/molecule/FileInputWithImage';
 import Button from 'components/atoms/Button';
-import { UserType } from 'types/userType';
-import { ProfileFormType } from 'types/profileFormType';
+import { UserType, ProfileFormType } from 'types';
 import { validateNickname } from 'utils/validatorUtils';
 
 interface Props {
