+++ frontend/src/pages/Setting2FAPage.tsx
@@ -4,7 +4,7 @@ import { deleteMy2fa, getMy2fa, getWhoami } from 'api/api.v1';
 import AppTemplate from 'components/templates/AppTemplate';
 import Setting2FA from 'components/organisms/Setting2FA';
 import HeaderWithBackButton from 'components/molecule/HeaderWithBackButton';
-import { UserType } from 'types/userType';
+import { UserType } from 'types';
 
 export default function Setting2FAPage() {
   const whoamiQuery = useQuery<UserType, AxiosError>({
