+++ frontend/src/pages/SettingBlocksPage.tsx
@@ -5,7 +5,7 @@ import AppTemplate from 'components/templates/AppTemplate';
 import HeaderWithBackButton from 'components/molecule/HeaderWithBackButton';
 import UserList from 'components/molecule/UserList';
 import Button from 'components/atoms/Button';
-import { UserType } from 'types/userType';
+import { UserType } from 'types';
 import styles from 'assets/styles/Blocks.module.css';
 
 export default function SettingBlocksPage() {
