+++ frontend/src/tests/ChannelBrowse.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import ChannelBrowse from 'components/organisms/ChannelBrowse';
 import { BrowserRouter } from 'react-router-dom';
-import { ChannelType } from 'types/channelType';
+import { ChannelType } from 'types';
 
 const dummyChannels: ChannelType[] = [
   {
