--- frontend/src/types/achievementType.ts
@@ -1,5 +1,0 @@
-export interface AchievementType {
-  id: string;
-  title: string;
-  description: string;
-}
