--- frontend/src/types/profileFormType.ts
@@ -1,4 +1,0 @@
-export interface ProfileFormType {
-  nickname: string;
-  imageFile: File | null;
-}
