--- frontend/src/types/userType.ts
@@ -1,12 +1,0 @@
-type UserStatus = 'online' | 'offline' | 'game';
-
-export interface UserType {
-  id: number;
-  nickname?: string;
-  rank?: number;
-  isTwoFactor?: boolean;
-  createdAt?: string;
-  updatedAt?: string;
-  deletedAt?: string;
-  status?: UserStatus;
-}
